requirements.txt
Flask==2.3.3
python-barcode[images]==0.15.1
